exported all covariance matrix estimation related functions from the ERGM class to utils; wrapped them with @njit; unified covariance matrix estimations such that now a single estimation with a single method is calculated at each optimization step, and it's used both as an hessian and for the convergence statistical test
