+++ tests/test_ergm.py
@@ -190,213 +190,3 @@ class TestERGM(unittest.TestCase):
 
         # TODO - what criteria to use for testing convergence? From manual tests, it doesn't seem to perfectly converge on the true thetas...
         # Nevertheless, even without an assert, this will catch errors - fit won't work if something breaks.
-
-    def test_auto_correlation_function(self):
-        n = 3
-        sample_size = 3
-        W1 = np.array([[0, 1, 0],
-                       [0, 0, 1],
-                       [1, 1, 0]])
-
-        W2 = np.array([[0, 1, 0],
-                       [0, 0, 1],
-                       [0, 1, 0]])
-
-        W3 = np.array([[0, 1, 1],
-                       [0, 0, 1],
-                       [1, 1, 0]])
-
-        sample = np.zeros((n, n, sample_size))
-        sample[:, :, 0] = W1
-        sample[:, :, 1] = W2
-        sample[:, :, 2] = W3
-
-        expected_gamma_hat_0 = 1 / 9 * np.array([[6, 3], [3, 2]])
-        expected_gamma_hat_1 = 1 / 27 * np.array([[-9, -6], [0, -1]])
-        expected_gamma_hat_2 = 1 / 27 * np.array([[0, 0], [-3, -2]])
-
-        ergm = ERGM(n, [NumberOfEdgesDirected(), TotalReciprocity()], True)
-
-        features_of_sample = ergm._network_statistics.calculate_sample_statistics(sample)
-        gammas = ergm.approximate_auto_correlation_function(features_of_sample)
-
-        self.assertTrue(np.abs(expected_gamma_hat_0 - gammas[0]).max() < 10 ** -15)
-        self.assertTrue(np.abs(expected_gamma_hat_1 - gammas[1]).max() < 10 ** -15)
-        self.assertTrue(np.abs(expected_gamma_hat_2 - gammas[2]).max() < 10 ** -15)
-
-    def test_capital_gammas(self):
-        n = 3
-        sample_size = 4
-
-        W1 = np.array([[0, 0, 1],
-                       [0, 0, 1],
-                       [1, 1, 0]])
-
-        W2 = np.array([[0, 0, 1],
-                       [0, 0, 1],
-                       [1, 1, 0]])
-
-        W3 = np.array([[0, 1, 0],
-                       [1, 0, 0],
-                       [0, 1, 0]])
-
-        W4 = np.array([[0, 0, 1],
-                       [1, 0, 0],
-                       [1, 0, 0]])
-
-        sample = np.zeros((n, n, sample_size))
-        sample[:, :, 0] = W1
-        sample[:, :, 1] = W2
-        sample[:, :, 2] = W3
-        sample[:, :, 3] = W4
-
-        expected_capital_gamma_0 = 5 / 16 * np.ones((2, 2))
-        expected_capital_gamma_1 = -3 / 16 * np.ones((2, 2))
-
-        ergm = ERGM(n, [NumberOfEdgesDirected(), TotalReciprocity()], True)
-        features_of_sample = ergm._network_statistics.calculate_sample_statistics(sample)
-        gammas = ergm.approximate_auto_correlation_function(features_of_sample)
-        capital_gammas = ergm.calc_capital_gammas(gammas)
-
-        self.assertTrue(np.abs(capital_gammas[0] - expected_capital_gamma_0).max() < 10 ** -15)
-        self.assertTrue(np.abs(capital_gammas[1] - expected_capital_gamma_1).max() < 10 ** -15)
-
-    @staticmethod
-    def _construct_sample():
-        n = 3
-        sample_size = 6
-        W1 = np.array([[0, 1, 0],
-                       [0, 0, 1],
-                       [1, 1, 0]])
-
-        W2 = np.array([[0, 1, 0],
-                       [1, 0, 0],
-                       [0, 1, 0]])
-
-        W3 = np.array([[0, 0, 1],
-                       [0, 0, 1],
-                       [1, 1, 0]])
-
-        W4 = np.array([[0, 0, 1],
-                       [1, 0, 1],
-                       [1, 1, 0]])
-
-        W5 = np.array([[0, 0, 1],
-                       [1, 0, 0],
-                       [1, 0, 0]])
-
-        W6 = np.array([[0, 0, 1],
-                       [1, 0, 1],
-                       [1, 0, 0]])
-
-        sample = np.zeros((n, n, sample_size))
-        sample[:, :, 0] = W1
-        sample[:, :, 1] = W2
-        sample[:, :, 2] = W3
-        sample[:, :, 3] = W4
-        sample[:, :, 4] = W5
-        sample[:, :, 5] = W6
-
-        return sample
-
-    def test_batches_covariance_estimation(self):
-        sample = self._construct_sample()
-        n = sample.shape[0]
-
-        expected_covariance_batch_estimation = np.ones((2, 2)) * 4 / 9
-
-        ergm = ERGM(n, [NumberOfEdgesDirected(), TotalReciprocity()], True)
-        features_of_sample = ergm._network_statistics.calculate_sample_statistics(sample)
-
-        batch_estimation = ergm.covariance_matrix_estimation(features_of_sample, method='batch', num_batches=3)
-        self.assertTrue(np.abs(expected_covariance_batch_estimation - batch_estimation).max() < 10 ** -15)
-
-    def test_naive_covariance_matrix_estimation(self):
-        sample = self._construct_sample()
-        n = sample.shape[0]
-        expected_covariance_naive_estimation = np.array([[17 / 36, 8 / 36], [8 / 36, 8 / 36]])
-
-        ergm = ERGM(n, [NumberOfEdgesDirected(), TotalReciprocity()], True)
-        features_of_sample = ergm._network_statistics.calculate_sample_statistics(sample)
-
-        naive_estimation = ergm.covariance_matrix_estimation(features_of_sample, method='naive')
-        self.assertTrue(np.abs(expected_covariance_naive_estimation - naive_estimation).max() < 10 ** -14)
-
-    @staticmethod
-    def validate_test_sample_validity(model, sample):
-        features_of_sample = model._network_statistics.calculate_sample_statistics(sample)
-        gamma_hats = ERGM.approximate_auto_correlation_function(features_of_sample)
-        capital_gammas = ERGM.calc_capital_gammas(gamma_hats)
-        Sigma_0 = -gamma_hats[0] + 2 * capital_gammas[0]
-        Sigma_1 = Sigma_0 + 2 * capital_gammas[1]
-        Sigma_2 = Sigma_1 + 2 * capital_gammas[2]
-
-        is_Sigma_0_positive_definite = eigh(Sigma_0)[0].min() > 0
-        is_Sigma_1_positive_definite = eigh(Sigma_1)[0].min() > 0
-        is_Sigma_2_positive_definite = eigh(Sigma_2)[0].min() > 0
-
-        det_Sigma_0 = np.linalg.det(Sigma_0)
-        det_Sigma_1 = np.linalg.det(Sigma_1)
-        det_Sigma_2 = np.linalg.det(Sigma_2)
-
-        if ((is_Sigma_0_positive_definite and det_Sigma_0 < det_Sigma_1) or
-                (is_Sigma_1_positive_definite and det_Sigma_1 < det_Sigma_2)):
-            return True
-        return False
-
-    @staticmethod
-    def find_sample_for_multivariate_initial_seq_covariance_matrix_estimation_test(num_trials=1000, sample_size=6, n=3,
-                                                                                   p=0.5):
-        ergm = ERGM(n, [NumberOfEdgesDirected(), TotalReciprocity()], True)
-        for i in range(num_trials):
-            sample = np.zeros((n, n, sample_size))
-            for j in range(sample_size):
-                cur_mat = nx.to_numpy_array(nx.erdos_renyi_graph(n, p, directed=True))
-                sample[:, :, j] = cur_mat
-                if TestERGM.validate_test_sample_validity(ergm, sample):
-                    return sample
-        print("No sample was valid")
-        return None
-
-    def test_multivariate_initial_seq_covariance_matrix_estimation(self):
-        n = 3
-        sample_size = 6
-        W1 = np.array([[0, 0, 0],
-                       [0, 0, 1],
-                       [1, 0, 0]])
-        W2 = np.array([[0, 1, 0],
-                       [0, 0, 0],
-                       [1, 1, 0]])
-        W3 = np.array([[0, 1, 1],
-                       [0, 0, 1],
-                       [1, 0, 0]])
-        W4 = np.array([[0, 1, 0],
-                       [0, 0, 0],
-                       [0, 1, 0]])
-        W5 = np.array([[0, 1, 0],
-                       [0, 0, 1],
-                       [1, 1, 0]])
-        W6 = np.array([[0, 1, 1],
-                       [1, 0, 0],
-                       [1, 1, 0]])
-        sample = np.zeros((n, n, sample_size))
-        sample[:, :, 0] = W1
-        sample[:, :, 1] = W2
-        sample[:, :, 2] = W3
-        sample[:, :, 3] = W4
-        sample[:, :, 4] = W5
-        sample[:, :, 5] = W6
-
-        # NOTE! this matrix was calculated after automatically calculating the capital gammas (using the dedicated
-        # function that has a dedicated test). I gave up on calculating everything from scratch after a long battle.
-        # Maybe we should come to that in the future.
-        expected_covariance_multivariate_initial_seq_estimation = 1 / (6 * 36) * np.array([[264, 192], [192, 144]])
-
-        ergm = ERGM(n, [NumberOfEdgesDirected(), TotalReciprocity()], True)
-        features_of_sample = ergm._network_statistics.calculate_sample_statistics(sample)
-
-        multivariate_initial_seq_estimation = ergm.covariance_matrix_estimation(features_of_sample,
-                                                                                method='multivariate_initial_sequence')
-        self.assertTrue(np.abs(
-            expected_covariance_multivariate_initial_seq_estimation - multivariate_initial_seq_estimation).max() <
-                        10 ** -14)
